Corrected Physics layers, fixed event logic,
One of the platforms in lvl 1 had the wrong physics layer. Many of the classes were using fixed update which is just for physics instead of just update, I corrected that. Some boolean logic in the bosses logic needed correcting so when he died he would actualy load the final scene. I added the audio manager to the credits scene.